Changed everything from float to double for consistency

+++ src/ACInside.java
@@ -38,7 +38,7 @@ public class ACInside
 
   // Talks to bank to place hold for a particular agent with a particular bidding key
   // Returns true if the placeHold was successful (i.e. the agent has the $ and is a valid person)
-  public boolean placeHold(Agent agent, float amount)
+  public boolean placeHold(Agent agent, Double amount)
   {
     //if(agent has money and is valid)
     //return true
